requirements.txt
# ======================
# Django + REST
# ======================
Django==5.0.2
djangorestframework==3.14.0
django-environ==0.11.2

# ======================
# PDF + OCR
# ======================
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
pdfplumber==0.10.3

# ======================
# Vector DB — Chroma
# (Render uses disk persistence)
# ======================
chromadb==0.4.24

# ======================
# ETL (DLT)
# ======================
dlt==0.4.10

# ======================
# Prefect Orchestration
# ======================
prefect==3.0.0

# ======================
# Logging
# ======================
loguru==0.7.2

# ======================
# Database drivers
# Use ONLY psycopg2-binary for Django + Postgres on Render
# ======================
psycopg2-binary==2.9.9

# ======================
# Environment Variables
# ======================
python-dotenv==1.0.1

# ======================
# Deployment (Render)
# ======================
gunicorn==21.2.0

# ======================
# Testing
# ======================
pytest==9.0.1
pytest-django==4.11.1
pytest-cov==7.0.0
